Capabilities/AgentsLangChainDashboard/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
pyyaml>=6.0
rich>=13.0.0
